+++ src/panels/lovelace/common/create-card-element.js
@@ -1,12 +1,12 @@
 import fireEvent from '../../../common/dom/fire_event.js';
 
 import '../cards/hui-camera-preview-card.js';
-import '../cards/hui-column-card.js';
 import '../cards/hui-entities-card.js';
 import '../cards/hui-entity-filter-card.js';
 import '../cards/hui-error-card.js';
 import '../cards/hui-glance-card';
 import '../cards/hui-history-graph-card.js';
+import '../cards/hui-horizontal-stack-card.js';
 import '../cards/hui-iframe-card.js';
 import '../cards/hui-markdown-card.js';
 import '../cards/hui-media-control-card.js';
@@ -14,19 +14,19 @@ import '../cards/hui-picture-elements-card';
 import '../cards/hui-picture-entity-card';
 import '../cards/hui-picture-glance-card';
 import '../cards/hui-plant-status-card.js';
-import '../cards/hui-row-card.js';
+import '../cards/hui-vertical-stack-card.js';
 import '../cards/hui-weather-forecast-card';
 
 import createErrorCardConfig from './create-error-card-config.js';
 
 const CARD_TYPES = [
   'camera-preview',
-  'column',
   'entities',
   'entity-filter',
   'error',
   'glance',
   'history-graph',
+  'horizontal-stack',
   'iframe',
   'markdown',
   'media-control',
@@ -34,7 +34,7 @@ const CARD_TYPES = [
   'picture-entity',
   'picture-glance',
   'plant-status',
-  'row',
+  'vertical-stack',
   'weather-forecast'
 ];
